requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.0.0
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
safetensors==0.4.1
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
peft==0.7.1
